Added overall_footer and overall_header to themes (#53)
It makes things a little messy because there is still foot.php which is shown in Vanilla but not People, and there is head.php which is shown on every page.
This was an unintuitive system to start with but I'm not sure how I want to deal with it for backwards-compatibility, so for now it can be a bit messy.